+++ back/brm-core/internal/model/contact.go
@@ -3,9 +3,9 @@ package model
 import "time"
 
 type Contact struct {
-	Id           uint
-	OwnerId      uint
-	EmployeeId   uint
+	Id           uint64
+	OwnerId      uint64
+	EmployeeId   uint64
 	Notes        string
 	CreationDate time.Time
 	IsDeleted    bool
@@ -17,6 +17,6 @@ type UpdateContact struct {
 }
 
 type GetContacts struct {
-	Limit  int
-	Offset int
+	Limit  uint
+	Offset uint
 }
