+++ back/registration/internal/model/company.go
@@ -1,11 +1,11 @@
 package model
 
 type Company struct {
-	Id           uint
+	Id           uint64
 	Name         string
 	Description  string
-	Industry     uint
-	OwnerId      uint
+	Industry     uint64
+	OwnerId      uint64
 	Rating       float64
 	CreationDate int64
 	IsDeleted    bool
